requirements.txt
pulp==2.6
Owlready2==0.36
pystardog==0.11.0
fastapi==0.75.0
python-decouple==3.6
uuid==1.30
uvicorn==0.15.0